Move notification permission to NearbyHelpFragment and ask for location permission less often
